+++ src/Dimensions/dimension.jl
@@ -294,36 +294,54 @@ Base.size(dims::DimTuple) = map(length, dims)
 Base.CartesianIndices(dims::DimTuple) = CartesianIndices(map(d -> axes(d, 1), dims))
 
 # Extents.jl
-function Extents.extent(ds::DimTuple, args...)
-    extent_dims = _astuple(dims(ds, args...))
-    extent_bounds = bounds(extent_dims)
-    return Extents.Extent{name(extent_dims)}(extent_bounds)
-end
-
-function _experimental_extent(ds::DimTuple)
-    regulardims = dims(ds, x -> !(lookup(x) isa MultiDimensionalLookup))    
-    regular_bounds = bounds.(regulardims)
+#=
+The complexity in this function is due to the fact that some lookups
+may have internal dimensions (see [hasinternaldimensions](@ref)).
+=#
+function Extents.extent(ids::DimTuple, args...)
+    ds = _astuple(dims(ids, args...))
+    # Obtain all dims which do _not_ have internal dimensions
+    regulardims = dims(ds, x -> !hasinternaldimensions(lookup(x)))    
+    # Their bounds are simple to obtain from `bounds`
+    regular_bounds = map(bounds, regulardims)
     regular_bounds_nt = NamedTuple{map(name, regulardims)}(regular_bounds)
     
+    # Now, for the dimensions which do have internal dimensions, we need to
+    # obtain the bounds for each of the internal dimensions.
     multidims = otherdims(ds, regulardims)
-    multidim_raw_bounds = bounds.(multidims) # we trust that bounds will give us a tuple of bounds one for each enclosed dimension
+    # Then, we get the `bounds` for each of the complex / higher order dimensions.
+    # `bounds` here returns an `ndims`-tuple of 2-tuples, unlike a simple 2-tuple for a regular dimension.
+    multidim_raw_bounds = map(bounds, multidims) # we trust that bounds will give us a tuple of bounds one for each enclosed dimension
+    # This gets us the list of "base dimensions" for each of the lookups.
     multidim_dims = combinedims(map(dims, multidims)...; length = false)
+    # We then determine the "true" bounds for each of the base dimensions in `multidim_dims`,
+    # knowing that some lookups might "overlap" in their internal dimensions.
+    # For each base dimension,
     multidim_bounds = map(multidim_dims) do outdim
-        foldl(zip(multidims, multidim_raw_bounds); init = (nothing, nothing)) do (minval, maxval), (dim, bounds)
+        # loop over the tuples of (multidim, multidim_raw_bounds) and reduce those to a single 2-tuple.
+        foldl(map(tuple, multidims, multidim_raw_bounds); init = (nothing, nothing)) do (minval, maxval), (dim, bounds)
+            # If the current dim "contains" the `outdim` we are interested in, then:
             if hasdim(dim, outdim)
                 if isnothing(minval) && isnothing(maxval)
+                    # Initialize the bounds to the bounds of the current dimension,
+                    # if no previous lookup has contained the `outdim`.
                     bounds[dimnum(dim, outdim)]
                 else
+                    # Otherwise, update the bounds to the minimum and maximum of the current and previous bounds.
                     this_dim_bounds = bounds[dimnum(dim, outdim)]
                     (min(minval, this_dim_bounds[1]), max(maxval, this_dim_bounds[2]))
                 end
             else
+                # Equivalent of a no-op, because the current dim does not contain the `outdim` we are interested in.
                 (minval, maxval)
             end
         end
     end
+    # Finally, create a NamedTuple of the bounds for each of the base dimensions,
+    # as Extents.jl expects.
     multidim_bounds_nt = NamedTuple{map(name, multidim_dims)}(multidim_bounds)
-    return merge(regular_bounds_nt, multidim_bounds_nt)
+    # Return the union of the base dimensions and the multidim dimensions.
+    return Extents.Extent(merge(regular_bounds_nt, multidim_bounds_nt))
 end
 
 dims(extent::Extents.Extent{K}) where K = map(rebuild, name2dim(K), values(extent))