+++ test/runtests.jl
@@ -22,6 +22,7 @@ end
 @time @safetestset "lookup" begin include("lookup.jl") end
 @time @safetestset "selector" begin include("selector.jl") end
 @time @safetestset "merged" begin include("merged.jl") end
+@time @safetestset "multidimensional_lookups" begin include("multidimensional_lookups.jl") end
 @time @safetestset "DimUnitRange" begin include("dimunitrange.jl") end
 @time @safetestset "format" begin include("format.jl") end
 @time @safetestset "array" begin include("array.jl") end